requirements.txt
python-dotenv
fastapi
uvicorn
langgraph
langchain
pydantic
flask
langchain-google-genai
python-docx
PyPDF2
gunicorn
